rofi-app: Factor out vtk stuff from Qt app to a separate header and cpp file
