+++ RoFILib/visualizer/CMakeLists.txt
@@ -12,7 +12,6 @@ target_link_libraries(rofi-vis-test PUBLIC configuration atoms ${ARMADILLO_LIBRA
 target_include_directories(rofi-vis-test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 
 
-
 find_package(Qt5 COMPONENTS Widgets REQUIRED)
 
 if(Qt5_FOUND)
@@ -22,7 +21,7 @@ if(Qt5_FOUND)
   set(CMAKE_AUTOUIC ON)
 
   add_executable(rofi-app
-    rofiapp_mainWindow.cpp rofiapp.cpp rofiapp_resources.qrc Camera.h Visualizer.h ${modelResources} )
+    rofiapp_mainWindow.cpp rofiapp.cpp rofiapp_resources.qrc Vtk.cpp ${modelResources} )
 
   target_link_libraries(rofi-app
     Qt5::Widgets configuration atoms cxxopts ${VTK_LIBRARIES})
