+++ RoFILib/visualizer/rofiapp_mainWindow.cpp
@@ -1,33 +1,6 @@
 #include "rofiapp_mainWindow.h"
 #include "ui_rofiapp_mainWindow.h"
 
-Matrix shoeMatrix()
-{
-    return rotate(M_PI/2, X);
-}
-
-Matrix bodyMatrix(double alpha)
-{
-    double diff = alpha * M_PI/180.0;
-    return rotate(M_PI/2 + diff, X);
-}
-
-Matrix dockMatrix(ConnectorId dock, bool on, double onCoeff = -1)
-{
-    double d;
-    if (onCoeff < 0){
-        d = on ? 0.05 : 0;
-    } else  {
-        d = onCoeff * 0.05;
-    }
-    Matrix docks[3] = {
-            translate(Vector{d,0,0}) * rotate(M_PI, Z), // XPlus
-            translate(Vector{-d,0,0}) * identity, // XMinus
-            translate(Vector{0,0,-d}) * rotate(-M_PI/2, Y) // ZMinus
-    };
-    return docks[dock];
-}
-
 Rofiapp_MainWindow::Rofiapp_MainWindow(QWidget *parent) :
     QMainWindow(parent),
     ui(new Ui::Rofiapp_MainWindow)
@@ -38,13 +11,6 @@ Rofiapp_MainWindow::Rofiapp_MainWindow(QWidget *parent) :
 
     current_cfg = new Configuration();
 
-    sphereSource = vtkSmartPointer<vtkSphereSource>::New();
-    sphereMapper = vtkSmartPointer<vtkPolyDataMapper>::New();
-    sphereMapper->SetInputConnection( sphereSource->GetOutputPort() );
-
-    sphereActor = vtkSmartPointer<vtkActor>::New();
-    sphereActor->GetProperty()->SetFrontfaceCulling(true);
-    sphereActor->SetMapper( sphereMapper );
 
     /* Fresh Renderer */
     renderWindow = vtkSmartPointer<vtkRenderWindow>::New();
@@ -57,8 +23,6 @@ Rofiapp_MainWindow::Rofiapp_MainWindow(QWidget *parent) :
 
     vtkSmartPointer<vtkCamera> camera = vtkSmartPointer<vtkCamera>::New();
     renderer->SetActiveCamera(camera);
-
-    addedActorsList = new QList<vtkSmartPointer<vtkActor>>;
 
     connect(ui->actionShow_sphere, SIGNAL(triggered()), this, SLOT(showSphere()));
     connect(ui->actionToggle_full_screen, SIGNAL(triggered()), this, SLOT(toggleFullScreen()));
@@ -74,12 +38,18 @@ Rofiapp_MainWindow::~Rofiapp_MainWindow()
 
 void Rofiapp_MainWindow::showSphere()
 {
-    for (QList<vtkSmartPointer<vtkActor>>::iterator i = addedActorsList->begin(); i != addedActorsList->end(); i++) {
-        renderer->RemoveActor( *i );
-    }
-    addedActorsList->clear();
+    /* Sphere */
+    vtkSmartPointer<vtkSphereSource> sphereSource = vtkSmartPointer<vtkSphereSource>::New();
+    vtkSmartPointer<vtkPolyDataMapper> sphereMapper = vtkSmartPointer<vtkPolyDataMapper>::New();
+    vtkSmartPointer<vtkActor> sphereActor = vtkSmartPointer<vtkActor>::New();
+
+    sphereMapper->SetInputConnection( sphereSource->GetOutputPort() );
+    sphereActor->GetProperty()->SetFrontfaceCulling(true);
+    sphereActor->SetMapper( sphereMapper );
+
+    renderer->RemoveAllViewProps();
+
     renderer->AddActor( sphereActor );
-    addedActorsList->append( sphereActor );
     renderer->ResetCamera();
     ui->qvtkWidget->update();
     ui->resetCamera->setEnabled(false);
@@ -131,75 +101,6 @@ void Rofiapp_MainWindow::on_loadConf_clicked()
     }
  }
 
-inline vtkSmartPointer<vtkMatrix4x4> convertMatrix( const Matrix& m )
-{
-    vtkSmartPointer<vtkMatrix4x4> mat = vtkSmartPointer<vtkMatrix4x4>::New();
-    mat->SetElement(0,0, m(0,0));
-    mat->SetElement(0,1, m(0,1));
-    mat->SetElement(0,2, m(0,2));
-    mat->SetElement(0,3, m(0,3));
-    mat->SetElement(1,0, m(1,0));
-    mat->SetElement(1,1, m(1,1));
-    mat->SetElement(1,2, m(1,2));
-    mat->SetElement(1,3, m(1,3));
-    mat->SetElement(2,0, m(2,0));
-    mat->SetElement(2,1, m(2,1));
-    mat->SetElement(2,2, m(2,2));
-    mat->SetElement(2,3, m(2,3));
-    mat->SetElement(3,0, m(3,0));
-    mat->SetElement(3,1, m(3,1));
-    mat->SetElement(3,2, m(3,2));
-    mat->SetElement(3,3, m(3,3));
-    return mat;
-}
-
-std::filesystem::path Rofiapp_MainWindow::getModel( const std::string& model ) {
-    static ResourceFile body = LOAD_RESOURCE_FILE( visualizer_model_body_obj );
-    static ResourceFile shoe = LOAD_RESOURCE_FILE( visualizer_model_shoe_obj );
-    static ResourceFile connector = LOAD_RESOURCE_FILE( visualizer_model_connector_obj );
-
-    if ( model == "body" )
-        return body.name();
-    if ( model == "shoe" )
-        return shoe.name();
-    if ( model == "connector" )
-        return connector.name();
-    throw std::runtime_error( "Invalid model '" + model + "' requested" );
-}
-
-void Rofiapp_MainWindow::addActor(const std::string &model, const Matrix &matrix, int color)
-{
-
-    // vtkOBJReader is locale sensitive!
-    std::locale currentLocale;
-    currentLocale = std::locale::global(std::locale::classic());
-    vtkSmartPointer<vtkOBJReader> reader = vtkSmartPointer<vtkOBJReader>::New();
-    reader->SetFileName( getModel(model).c_str() );
-    reader->Update();
-    std::locale::global(currentLocale);
-
-    vtkSmartPointer<vtkTransform> rotation = vtkSmartPointer<vtkTransform>::New();
-    rotation->SetMatrix( convertMatrix(matrix) );
-
-    vtkSmartPointer<vtkTransformPolyDataFilter> filter = vtkSmartPointer<vtkTransformPolyDataFilter>::New();
-    filter->SetTransform( rotation );
-    filter->SetInputConnection( reader->GetOutputPort() );
-
-    vtkSmartPointer<vtkPolyDataMapper> frameMapper = vtkSmartPointer<vtkPolyDataMapper>::New();
-    frameMapper->SetInputConnection(filter->GetOutputPort());
-
-    vtkSmartPointer<vtkActor> frameActor = vtkSmartPointer<vtkActor>::New();
-    frameActor->SetMapper(frameMapper);
-    frameActor->GetProperty()->SetColor(colors[color][0]/256.0, colors[color][1]/256.0 , colors[color][2]/256.0);
-    frameActor->GetProperty()->SetOpacity(1.0);
-    frameActor->GetProperty()->SetFrontfaceCulling(true);
-    frameActor->GetProperty()->SetBackfaceCulling(true);
-    frameActor->SetPosition( matrix(0,3), matrix(1,3), matrix(2,3) );
-    frameActor->SetScale( 1 / 95.0 );
-
-    renderer->AddActor( frameActor );
-    addedActorsList->append( frameActor );
-}
 
 void Rofiapp_MainWindow::on_showConf_clicked()
 {
@@ -210,44 +111,17 @@ void Rofiapp_MainWindow::on_showConf_clicked()
     current_cfg = new Configuration();
 
     std::istringstream iStream(lineContents.toStdString());
-    if (!IO::readConfiguration(iStream, *current_cfg)) {
-        return;
+    try { IO::readConfiguration(iStream, *current_cfg); }
+    catch (...) {
+        qDebug("Error on parsing configuration.");
     }
+    current_cfg -> computeMatrices();
 
-    for (QList<vtkSmartPointer<vtkActor>>::iterator i = addedActorsList->begin(); i != addedActorsList->end(); i++) {
-        renderer->RemoveActor( *i );
-    }
-    addedActorsList->clear();
+    renderer->RemoveAllViewProps();
+    VtkSupp::buildScene(current_cfg, renderer);
 
-    current_cfg->computeMatrices();
-    for ( const auto& [id, matrices] : current_cfg->getMatrices())
-    {
-        int color =  id % 7 + 3;
-        const Module& mod = current_cfg->getModules().at(id);
-        EdgeList edges = current_cfg->getEdges().at(id);
-        for (ShoeId s : {A, B})
-        {
-            Joint j = s == A ? Alpha : Beta;
-            this->addActor("shoe", matrices[s] * shoeMatrix(), color);
-            this->addActor("body", matrices[s] * bodyMatrix(mod.getJoint(j)), color);
-
-            for (ConnectorId dock : {XPlus, XMinus, ZMinus})
-            {
-                bool on = edges[s * 3 + dock].has_value();
-                double onCoeff = on ? edges[s * 3 + dock].value().onCoeff() : 0;
-                this->addActor("connector", matrices[s] * dockMatrix(dock, on, onCoeff), color);
-            }
-        }
-    }
-
-    camera = renderer -> GetActiveCamera();
-    Vector massCenter = current_cfg->massCenter();
-    camera->SetFocalPoint(massCenter(0), massCenter(1), massCenter(2));
-    camera->SetPosition(massCenter(0), massCenter(1) - 6, massCenter(2));
-    camera->SetViewUp(0,0,1);
     ui->resetCamera->setEnabled(true);
-
-    ui->qvtkWidget->update();
+    this -> on_resetCamera_clicked();
 }
 
 void Rofiapp_MainWindow::on_resetCamera_clicked()
@@ -257,8 +131,8 @@ void Rofiapp_MainWindow::on_resetCamera_clicked()
     camera->SetFocalPoint(massCenter(0), massCenter(1), massCenter(2));
     camera->SetPosition(massCenter(0), massCenter(1) - 6, massCenter(2));
     camera->SetViewUp(0,0,1);
-    ui->qvtkWidget->update();
 
+    ui->qvtkWidget->update();
 }
 
 void Rofiapp_MainWindow::on_configTextWindow_textChanged()
